src: stats: Refactor accumulated part into its own module
